+++ lib/glimmer/libui/shape.rb
@@ -23,6 +23,7 @@ require 'glimmer/libui/parent'
 require 'glimmer/libui/control_proxy/area_proxy'
 require 'glimmer/libui/control_proxy/path_proxy'
 require 'glimmer/libui/data_bindable'
+require 'glimmer/libui/perfect_shaped'
 
 module Glimmer
   module LibUI
@@ -65,6 +66,7 @@ module Glimmer
       end
       
       include Parent
+      include PerfectShaped
       include DataBindable
       
       attr_reader :parent, :args, :keyword, :block, :content_added
@@ -145,47 +147,7 @@ module Glimmer
       end
       alias transform= transform
       alias set_transform transform
-    
-      # Returns if shape contains point on the inside when outline is false (default)
-      # or if point is on the outline when outline is true
-      # distance_tolerance is used when outline is true to enable a fuzz factor in
-      # determining if a point lies on the outline (e.g. makes it easier to select
-      # a shape by mouse)
-      def contain?(*point, outline: false, distance_tolerance: 0)
-        perfect_shape&.contain?(*point, outline: outline, distance_tolerance: distance_tolerance)
-      end
-      
-      # Returns if shape includes point on the inside when filled
-      # or if shape includes point on the outline when stroked
-      def include?(*point)
-        if fill.empty?
-          contain?(*point, outline: true, distance_tolerance: ((stroke[:thickness] || 1) - 1))
-        else
-          contain?(*point)
-        end
-      end
-      
-      # Returns bounding box Array consisting of
-      # [minx, miny, width, height]
-      def bounding_box
-        perfect_shape_bounding_box = perfect_shape&.bounding_box
-        return if perfect_shape_bounding_box.nil?
-        [
-          perfect_shape_bounding_box.x,
-          perfect_shape_bounding_box.y,
-          perfect_shape_bounding_box.width,
-          perfect_shape_bounding_box.height,
-        ]
-      end
-      
-      # Returns PerfectShape object matching this shape to enable
-      # executing computational geometry algorithms
-      #
-      # Subclasses must implement
-      def perfect_shape
-        # No Op
-      end
-      
+
       def respond_to?(method_name, *args, &block)
         self.class.parameters.include?(method_name.to_s.sub(/=$/, '').sub(/^set_/, '').to_sym) or
           super(method_name, true)